Database version 15: Split out sticky links and group links.
The group_link table has been created so the circular dependency issues can be resolved. Additionally, the sticky_link and normal_link tables have been created instead of the link table with the style column. This makes it easier to build a smaller graph, since we don't need to select outgoing sticky links.