Added t3060 - new circular group test.
This one is more complicated than the first, since we need to keep track of the group dependencies in order to detect that writing to a group will cause a circular dependency at parsing time. Deferring the circular dependency check to the updater will not work, since the PDAG may not have all the groups loaded, and so an incremental build does not necessarily load a circular DAG. However, future updates may load the circular DAG and then fail. It is necessary to detect these errors before they are inserted into the database.

The detection works by building a PDAG using the group_link table from the group outputs of commands that have new group inputs or outputs. We have to defer building the PDAG until all Tupfiles have been parsed, since one Tupfile may remove a group link, while another Tupfile may add the inverse. (Eg: A/Tupfile has a link from <foo> to <bar>, then on the next parse A/Tupfile removes that group dependency while B/Tupfile adds a link from <bar> to <foo>. We have to parse both Tupfiles before checking for circular dependencies.) The circular portion of the group DAG is saved in a temporary file for debugging.